Add SciSharp.TensorFlow-Cpu.Redist nuget project.
This project can be used to generate a nuget holding tensorflow cpu
libraries for windows and linux runtimes.

The Copy-NativeTensorFlowLibs.ps1 script can be used to download all the
libraries from Google.

The examples have been updated to use this nuget to check that things
are working.
(Make sure to update your package source to have the packages directory
in your nuget source before trying this at home).